OCR/requirements.txt
chatllm
unstructured
pdf2image
pdfminer.six
IPython
unstructured_inference
rapidocr_onnxruntime
python-dotenv